Added directory specifying and error checking
Implemented an additional command line argument that allows for the
specification of a directory to output shares and the prime to, or pull
the shares and prime from. This defaults to the current working
directory if not specified.

Addtional error checking was added to cover for unusual argument values
that are not valid for secret sharing.